Neural Network model created and parameters truned
The neural network model to perform binary classification on the income dataset was created with 2 hidden layers and 16 nodes in each hidden layers.
Next, hyper parameter tuning was done and compared to base model to check accuracy differences.
Lastly, methods like EarlyStopping, K-Folds and ShuffleSplit was applied to improve the accuracy and show the accuracy comparison with the base model.